requirements.txt
numpy==2.1.*
pandas==2.2.*
pipx==1.7.*
streamlit==1.38.*
plotly==5.24.*
python-dotenv==1.0.*
rich==13.8.*
scikit-learn==1.5.*
lightgbm==4.5.*
xgboost==2.1.*
black==23.11.*
mypy==1.7.*
sklego
streamlit_option_menu
streamlit-pandas-profiling
